+++ src/parse_map/main.c
@@ -6,7 +6,7 @@
 /*   By: gmasid <gmasid@student.42.rio>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/27 18:19:23 by gmasid            #+#    #+#             */
-/*   Updated: 2023/03/31 21:30:09 by gmasid           ###   ########.fr       */
+/*   Updated: 2023/03/31 21:39:28 by gmasid           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,34 +42,6 @@ void	parse_map_file_and_set_textures(char *path, t_game *game)
 	if (!matrix_len(game->map.array))
 		exit_app("Map is empty", game);
 	close(fd);
-}
-
-void	normalize_map(t_game *game)
-{
-	char	**temp;
-	int		i;
-	int		j;
-
-	temp = malloc(sizeof(char *) * (game->map.height + 1));
-	i = 0;
-	while (i < game->map.height)
-	{
-		j = 0;
-		temp[i] = malloc(sizeof(char) * (game->map.width + 1));
-		while (j < game->map.width)
-		{
-			if (j >= (int)ft_strlen(game->map.array[i]))
-				temp[i][j] = ' ';
-			else
-				temp[i][j] = game->map.array[i][j];
-			j++;
-		}
-		temp[i][j] = '\0';
-		i++;
-	}
-	temp[i] = NULL;
-	free_matrix(game->map.array);
-	game->map.array = temp;
 }
 
 void	parse_map(char *path, t_game *game)
